fix(backend): migration constant, move mapping tm-211
