+++ apps/backend/src/modules/vendors/libs/enums/enums.ts
@@ -1,5 +1,5 @@
 export { EdxApiPath } from "./edx-api-path.enum.js";
-export { EdxCourseFieldsMapping } from "./edx-course-fields-mapping.enum.js";
+export { EdxCourseField } from "./edx-course-field.enum.js";
 export { UdemyApiPath } from "./udemy-api-path.enum.js";
 export { CourseField } from "./udemy-course-field.enum.js";
 export { UdemyCourseFieldsMapping } from "./udemy-course-fields-mapping.enum.js";
