+++ apps/backend/src/modules/vendors/libs/mapper/mapper.ts
@@ -1,0 +1,1 @@
+export { EdxCourseFieldsMapping } from "./edx-course-fields.mapper.js";
